refactor: update ProgramType imports and enhance type definitions across multiple files
