+++ libs/data/src/project/projectFeatures.ts
@@ -3,7 +3,7 @@ import { fileURLToPath } from 'url'
 import { ProjectBaserow } from '../common/baserow/projectBaserow'
 import { DataProject } from './types/domain'
 import { jsonPrograms } from '../../static'
-import { ProgramType } from '../index'
+import { ProgramType } from '../program/types/shared'
 import { ThemeId } from '../theme/types/shared'
 import { SlugValidator } from '../common/validators/slugValidator'
 import { LinkValidator } from '../common/validators/linkValidator'
