+++ libs/data/src/server.ts
@@ -1,2 +1,3 @@
 export * from './statistics/posthogUpdater'
 export * from './program/program'
+export * from './program/types/shared'
